+++ crates/holochain/src/conductor/conductor.rs
@@ -1660,10 +1660,7 @@ mod app_impls {
             self: Arc<Self>,
             payload: InstallAppPayload,
         ) -> ConductorResult<InstalledApp> {
-            #[cfg(feature = "chc")]
             let ignore_genesis_failure = payload.ignore_genesis_failure;
-            #[cfg(not(feature = "chc"))]
-            let ignore_genesis_failure = false;
 
             let InstallAppPayload {
                 source,
@@ -1902,7 +1899,8 @@ mod app_impls {
                 let installed_app_id = installed_app_id.clone();
                 move |mut state| {
                     let app = state.get_app_mut(&installed_app_id)?;
-                    app.status = AppStatus::Disabled(DisabledAppReason::NeverStarted);
+                    app.status =
+                        AppStatus::Disabled(DisabledAppReason::NotStartedAfterProvidingMemproofs);
                     Ok(state)
                 }
             })
@@ -1961,6 +1959,10 @@ mod app_impls {
 
 /// Methods related to cell access
 mod cell_impls {
+    use std::collections::BTreeSet;
+
+    use holochain_conductor_api::CompatibleCells;
+
     use super::*;
 
     impl Conductor {
@@ -1994,6 +1996,51 @@ mod cell_impls {
         pub fn running_cell_ids(&self) -> HashSet<CellId> {
             self.running_cells
                 .share_ref(|cells| cells.keys().cloned().collect())
+        }
+
+        /// Returns all installed cells which are forward compatible with the specified DNA,
+        /// including direct matches, by examining the "lineage" specified by DNAs of currently installed cells.
+        ///
+        /// Each DnaDef specifies a "lineage" field of DNA hashes, which indicates that the DNA is forward-compatible
+        /// with the DNAs specified in its lineage. If the DnaHash parameter is contained within the lineage of any
+        /// installed cell's DNA, that cell will be returned in the result set, since it has declared
+        /// itself forward-compatible.
+        pub async fn cells_by_dna_lineage(
+            &self,
+            dna_hash: &DnaHash,
+        ) -> ConductorResult<CompatibleCells> {
+            // TODO: OPTIMIZE: cache the DNA lineages
+            Ok(self
+                .get_state()
+                .await?
+                // Look in all installed apps
+                .installed_apps_and_services()
+                .values()
+                .filter_map(|app| {
+                    let cells_in_lineage: BTreeSet<_> = app
+                        // Look in all cells for the app
+                        .all_cells()
+                        .filter_map(|cell_id| {
+                            let cell_dna_hash = cell_id.dna_hash();
+                            if cell_dna_hash == dna_hash {
+                                // If a direct hit, include this CellId in the list of candidates
+                                Some(cell_id.clone())
+                            } else {
+                                // If this cell *contains* the given DNA in *its* lineage, include it.
+                                self.get_dna_def(cell_id.dna_hash())
+                                    .map(|dna_def| dna_def.lineage.contains(dna_hash))
+                                    .unwrap_or(false)
+                                    .then(|| cell_id.clone())
+                            }
+                        })
+                        .collect();
+                    if cells_in_lineage.is_empty() {
+                        None
+                    } else {
+                        Some((app.installed_app_id.clone(), cells_in_lineage))
+                    }
+                })
+                .collect())
         }
     }
 }
@@ -3161,7 +3208,10 @@ impl Conductor {
     }
 
     /// Remove all Cells which are not referenced by any Enabled app.
-    /// (Cells belonging to Paused apps are not considered "dangling" and will not be removed)
+    /// (Cells belonging to Paused apps are not considered "dangling" and will not be removed).
+    ///
+    /// Additionally, if the cell is being removed because the last app referencing it was uninstalled,
+    /// all data used by that cell (across Authored, DHT, and Cache databases) will also be removed.
     #[tracing::instrument(skip_all)]
     async fn remove_dangling_cells(&self) -> ConductorResult<()> {
         let state = self.get_state().await?;
@@ -3176,6 +3226,8 @@ impl Conductor {
             .iter()
             .flat_map(|(_, app)| app.all_cells().collect::<HashSet<_>>())
             .collect();
+
+        let all_dnas: HashSet<_> = all_cells.iter().map(|cell_id| cell_id.dna_hash()).collect();
 
         // Clean up all cells that will be dropped (leave network, etc.)
         let cells_to_cleanup: Vec<_> = self.running_cells.share_mut(|cells| {
@@ -3198,53 +3250,54 @@ impl Conductor {
             cell.cleanup().await?;
         }
 
+        // Find any cleaned up cells which are no longer used by any app,
+        // so that we can remove their data from the databases.
+        let cells_to_purge = cells_to_cleanup
+            .iter()
+            .filter_map(|cell| (!all_cells.contains(cell.id())).then_some(cell.id().clone()));
+
         // Find any DNAs from cleaned up cells which don't have representation in any cells
-        // in any app. In other words, find the DNAs which are *only* represented in uninstalled apps.
-        let all_dnas: HashSet<_> = all_cells
-            .into_iter()
-            .map(|cell_id| cell_id.dna_hash().clone())
-            .collect();
-        let dnas_to_cleanup = cells_to_cleanup
+        // in any installed app, so that we can remove their data from the databases.
+        let dnas_to_purge = cells_to_cleanup
             .iter()
             .map(|cell| cell.id().dna_hash())
             .filter(|dna| !all_dnas.contains(dna));
 
-        // For any unrepresented DNAs, clean up those DNA-specific databases
-        for dna_hash in dnas_to_cleanup {
-            futures::future::join_all(
-                self.spaces
-                    .get_all_authored_dbs(dna_hash)
-                    .unwrap()
-                    .iter()
-                    .map(|db| {
-                        db.write_async(|txn| -> DatabaseResult<usize> {
-                            Ok(txn.execute("DELETE FROM Action", ())?)
-                        })
-                        .instrument(tracing::info_span!("cleanup authored db", ?dna_hash))
-                        .boxed()
-                    }),
-            )
-            .await
-            .into_iter()
-            .collect::<Result<Vec<usize>, _>>()?;
+        // Delete all data from authored databases which are longer installed
+        for cell_id in cells_to_purge {
+            let db = self
+                .spaces
+                .get_or_create_authored_db(cell_id.dna_hash(), cell_id.agent_pubkey().clone())?;
+            let mut path = db.path().clone();
+            if let Err(err) = ffs::remove_file(&path).await {
+                tracing::warn!(?err, "Could not remove primary DB file, probably because it is still in use. Purging all data instead.");
+                db.write_async(purge_data).await?;
+            }
+            path.set_extension("");
+            let stem = path.to_string_lossy();
+            for ext in ["db-shm", "db-wal"] {
+                let path = PathBuf::from(format!("{stem}.{ext}"));
+                if let Err(err) = ffs::remove_file(&path).await {
+                    let err = err.remove_backtrace();
+                    tracing::warn!(?err, "Failed to remove DB file");
+                }
+            }
+        }
 
+        // For any DNAs no longer represented in any installed app,
+        // purge data from those DNA-specific databases
+        for dna_hash in dnas_to_purge {
             futures::future::join_all(
                 [
                     self.spaces
                         .dht_db(dna_hash)
                         .unwrap()
-                        .write_async(|txn| {
-                            DatabaseResult::Ok(txn.execute("DELETE FROM Action", ())?)
-                        })
-                        .in_current_span()
+                        .write_async(purge_data)
                         .boxed(),
                     self.spaces
                         .cache(dna_hash)
                         .unwrap()
-                        .write_async(|txn| {
-                            DatabaseResult::Ok(txn.execute("DELETE FROM Action", ())?)
-                        })
-                        .in_current_span()
+                        .write_async(purge_data)
                         .boxed(),
                     // TODO: also delete stale Wasms
                 ]
@@ -3252,7 +3305,7 @@ impl Conductor {
             )
             .await
             .into_iter()
-            .collect::<Result<Vec<usize>, _>>()?;
+            .collect::<Result<Vec<()>, _>>()?;
         }
 
         Ok(())
@@ -3556,6 +3609,17 @@ mod test_utils_impls {
             Ok(cell.triggers().clone())
         }
     }
+}
+
+#[tracing::instrument(skip_all)]
+fn purge_data(txn: &mut Transaction) -> DatabaseResult<()> {
+    txn.execute("DELETE FROM DhtOp", ())?;
+    txn.execute("DELETE FROM Action", ())?;
+    txn.execute("DELETE FROM Entry", ())?;
+    txn.execute("DELETE FROM ValidationReceipt", ())?;
+    txn.execute("DELETE FROM ChainLock", ())?;
+    txn.execute("DELETE FROM ScheduledFunctions", ())?;
+    Ok(())
 }
 
 /// Perform Genesis on the source chains for each of the specified CellIds.