+++ crates/holochain/src/conductor/conductor/tests.rs
@@ -297,12 +297,30 @@ async fn test_uninstall_app() {
         (2, 0)
     );
 
+    let db1 = conductor
+        .spaces
+        .get_or_create_authored_db(dna.dna_hash(), app1.cells()[0].agent_pubkey().clone())
+        .unwrap();
+    let db2 = conductor
+        .spaces
+        .get_or_create_authored_db(dna.dna_hash(), app2.cells()[0].agent_pubkey().clone())
+        .unwrap();
+
+    // - Check that both authored database files exist
+    std::fs::File::open(db1.path()).unwrap();
+    std::fs::File::open(db2.path()).unwrap();
+
     // - Uninstall the first app
     conductor
         .raw_handle()
         .uninstall_app(&"app1".to_string())
         .await
         .unwrap();
+
+    // - Check that the first authored DB file is deleted since the cell was removed.
+    #[cfg(not(windows))]
+    std::fs::File::open(db1.path()).unwrap_err();
+    std::fs::File::open(db2.path()).unwrap();
 
     // - Ensure that the remaining app can still access both hashes
     assert!(conductor
@@ -320,6 +338,10 @@ async fn test_uninstall_app() {
         .uninstall_app(&"app2".to_string())
         .await
         .unwrap();
+
+    // - Check that second authored DB file is deleted since the cell was removed.
+    #[cfg(not(windows))]
+    std::fs::File::open(db2.path()).unwrap_err();
 
     // - Ensure that the apps are removed
     assert_eq_retry_10s!(
@@ -1181,7 +1203,7 @@ async fn test_init_concurrency() {
 /// - conductor can be restarted and app still in AwaitingMemproofs state,
 /// - app functions normally after memproofs provided
 #[tokio::test(flavor = "multi_thread")]
-async fn test_deferred_provisioning() {
+async fn test_deferred_memproof_provisioning() {
     holochain_trace::test_run();
     let (dna, _, _) = SweetDnaFile::unique_from_test_wasms(vec![TestWasm::Foo]).await;
     let mut conductor = SweetConductor::from_standard_config().await;
@@ -1198,7 +1220,6 @@ async fn test_deferred_provisioning() {
             installed_app_id: Some(app_id.clone()),
             membrane_proofs: Default::default(),
             network_seed: None,
-            #[cfg(feature = "chc")]
             ignore_genesis_failure: false,
         })
         .await
@@ -1268,6 +1289,17 @@ async fn test_deferred_provisioning() {
         .await
         .unwrap();
 
+    //- Status is now Disabled with the special `NotStartedAfterProvidingMemproofs` reason.
+    //    It's not tested in this test, but this status allows the app to be enabled
+    //    over the app interface.
+    let app_info = conductor.get_app_info(&app_id).await.unwrap().unwrap();
+    assert_eq!(
+        app_info.status,
+        AppInfoStatus::Disabled {
+            reason: DisabledAppReason::NotStartedAfterProvidingMemproofs
+        }
+    );
+
     conductor.enable_app(app_id.clone()).await.unwrap();
 
     //- Status is now Running and there is 1 cell assignment
@@ -1280,7 +1312,7 @@ async fn test_deferred_provisioning() {
 
 /// Can uninstall an app with deferred memproofs before providing memproofs
 #[tokio::test(flavor = "multi_thread")]
-async fn test_deferred_provisioning_uninstall() {
+async fn test_deferred_memproof_provisioning_uninstall() {
     holochain_trace::test_run();
     let (dna, _, _) = SweetDnaFile::unique_from_test_wasms(vec![TestWasm::Foo]).await;
     let conductor = SweetConductor::from_standard_config().await;
@@ -1297,7 +1329,6 @@ async fn test_deferred_provisioning_uninstall() {
             installed_app_id: Some(app_id.clone()),
             membrane_proofs: Default::default(),
             network_seed: None,
-            #[cfg(feature = "chc")]
             ignore_genesis_failure: false,
         })
         .await