+++ crates/holochain/src/core/workflow/validation_receipt_workflow/tests.rs
@@ -157,14 +157,6 @@ async fn test_block_invalid_receipt() {
         ))?;
         Ok(hash)
     });
-    // .function(
-    //     coordinator_name,
-    //     get_function_name,
-    //     move |api, hash: AnyDhtHash| {
-    //         let records = api.get(vec![GetInput::new(hash, Default::default())])?;
-    //         Ok(records[0])
-    //     },
-    // );
 
     let zomes_that_check = InlineZomeSet::new_single(
         integrity_name,
@@ -215,9 +207,14 @@ async fn test_block_invalid_receipt() {
         .call(&alice_cell.zome(coordinator_name), create_function_name, ())
         .await;
 
-    await_consistency(10, [&alice_cell, &bob_cell])
-        .await
-        .unwrap();
+    // Don't check alice's integrated ops, since she gets blocked during gossip
+    await_consistency_advanced(
+        10,
+        vec![(alice_cell.agent_pubkey().clone(), 1)],
+        [(&alice_cell, false), (&bob_cell, true)],
+    )
+    .await
+    .unwrap();
 
     let alice_block_target = BlockTargetId::Cell(alice_cell.cell_id().to_owned());
     let bob_block_target = BlockTargetId::Cell(bob_cell.cell_id().to_owned());