+++ crates/holochain/tests/tests/test_utils/mod.rs
@@ -281,7 +281,6 @@ pub async fn register_and_install_dna_named(
         installed_app_id: Some(name),
         network_seed: None,
         membrane_proofs: std::collections::HashMap::new(),
-        #[cfg(feature = "chc")]
         ignore_genesis_failure: false,
     };
     let request = AdminRequest::InstallApp(Box::new(payload));